# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cipla Healthcare Campaign Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/managers:
    get:
      operationId: listManagers
      tags: [admin]
      summary: List all managers with their stats
      responses:
        "200":
          description: List of managers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManagerWithStats"
    post:
      operationId: createManager
      tags: [admin]
      summary: Create a new manager
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagerInput"
      responses:
        "201":
          description: Manager created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/managers/{id}:
    delete:
      operationId: deleteManager
      tags: [admin]
      summary: Delete a manager
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Manager deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /manager/doctors:
    get:
      operationId: listMyDoctors
      tags: [manager]
      summary: List doctors for the current manager
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: city
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated doctor list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorListResponse"
    post:
      operationId: createDoctor
      tags: [manager]
      summary: Add a new doctor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctorInput"
      responses:
        "201":
          description: Doctor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
        "409":
          description: Duplicate doctor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /manager/doctors/{id}:
    get:
      operationId: getDoctor
      tags: [manager]
      summary: Get a single doctor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Doctor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    User:
      type: object
      required: [id, name, employeeCode, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        employeeCode:
          type: string
        role:
          type: string
          enum: [admin, manager]
        createdAt:
          type: string
    ManagerInput:
      type: object
      required: [name, employeeCode, password]
      properties:
        name:
          type: string
        employeeCode:
          type: string
        password:
          type: string
    ManagerWithStats:
      type: object
      required: [id, name, employeeCode, totalDoctors, photosUploaded, documentsUploaded, completedProfiles, completionPercentage, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        employeeCode:
          type: string
        totalDoctors:
          type: integer
        photosUploaded:
          type: integer
        documentsUploaded:
          type: integer
        completedProfiles:
          type: integer
        completionPercentage:
          type: number
        createdAt:
          type: string
    Doctor:
      type: object
      required: [id, managerId, doctorName, specialization, city, createdAt, isComplete]
      properties:
        id:
          type: integer
        managerId:
          type: integer
        managerName:
          type: ["string", "null"]
        doctorName:
          type: string
        specialization:
          type: string
        city:
          type: string
        clinicAddress:
          type: ["string", "null"]
        phoneNumber:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        documentUrl:
          type: ["string", "null"]
        photoUploadedAt:
          type: ["string", "null"]
        documentUploadedAt:
          type: ["string", "null"]
        isComplete:
          type: boolean
        createdAt:
          type: string
    DoctorInput:
      type: object
      required: [doctorName, specialization, city]
      properties:
        doctorName:
          type: string
        specialization:
          type: string
        city:
          type: string
        clinicAddress:
          type: ["string", "null"]
        phoneNumber:
          type: ["string", "null"]
    DoctorListResponse:
      type: object
      required: [doctors, total, page, limit, totalPages]
      properties:
        doctors:
          type: array
          items:
            $ref: "#/components/schemas/Doctor"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
...