# lib/api-zod/src/generated/types/doctorInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cipla Healthcare Campaign Management API
 * OpenAPI spec version: 0.1.0
 */

export interface DoctorInput {
  doctorName: string;
  specialization: string;
  city: string;
  /** @nullable */
  clinicAddress?: string | null;
  /** @nullable */
  phoneNumber?: string | null;
}